% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{generate_keyword_index}
\alias{generate_keyword_index}
\title{generate keyword index}
\usage{
generate_keyword_index(df_in, column_value = "name",
  keyword = "dehalogenase")
}
\arguments{
\item{df_in}{is the input data.frame}

\item{column_value}{is the name of the column to select}

\item{is}{a key word to search in the column}
}
\description{
index a dataframe by searching for a keyword in a column of a dataframe
}

