% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{keyword_df}
\alias{keyword_df}
\title{keyword df}
\usage{
keyword_df(df_in, keyword = NULL)
}
\arguments{
\item{df_in}{}
}
\value{
df_out
}
\description{
adds a keyword to the dataframe as a new column
}

