% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{open_dataframe}
\alias{open_dataframe}
\title{open dataframe}
\usage{
open_dataframe(rds_file = "dehalogenators.rds")
}
\arguments{
\item{rds_file}{}
}
\value{
df a data.frame
}
\description{
reads and rds file to a dataframe
}

