% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{subset_df_by_row_ind}
\alias{subset_df_by_row_ind}
\title{subset df by row ind}
\usage{
subset_df_by_row_ind(df_in, ind)
}
\arguments{
\item{df_in}{}

\item{ind}{}
}
\value{
df_out
}
\description{
passes a logical index to subset df by row
}

